+++ packages/rum-core/src/boot/startRum.spec.ts
@@ -1,14 +1,16 @@
-import { RelativeTime, Configuration, Observable } from '@datadog/browser-core'
+import { RelativeTime, Configuration, Observable, noop, relativeNow } from '@datadog/browser-core'
 import { RumSession } from '@datadog/browser-rum-core'
 import { createRumSessionMock, RumSessionMock } from '../../test/mockRumSession'
 import { isIE } from '../../../core/test/specHelper'
 import { noopRecorderApi, setup, TestSetupBuilder } from '../../test/specHelper'
-import { RumPerformanceNavigationTiming } from '../browser/performanceCollection'
+import { RumPerformanceNavigationTiming, RumPerformanceEntry } from '../browser/performanceCollection'
 
 import { LifeCycle, LifeCycleEventType } from '../domain/lifeCycle'
 import { SESSION_KEEP_ALIVE_INTERVAL, THROTTLE_VIEW_UPDATE_PERIOD } from '../domain/rumEventsCollection/view/trackViews'
 import { startViewCollection } from '../domain/rumEventsCollection/view/viewCollection'
 import { RumEvent } from '../rumEvent.types'
+import { LocationChange } from '../browser/locationChangeObservable'
+import { startLongTaskCollection } from '../domain/rumEventsCollection/longTask/longTaskCollection'
 import { startRumEventCollection } from './startRum'
 
 function collectServerEvents(lifeCycle: LifeCycle) {
@@ -25,13 +27,16 @@ function startRum(
   configuration: Configuration,
   session: RumSession,
   location: Location,
-  domMutationObservable: Observable<void>
+  domMutationObservable: Observable<void>,
+  locationChangeObservable: Observable<LocationChange>
 ) {
   const { stop: rumEventCollectionStop, foregroundContexts } = startRumEventCollection(
     applicationId,
     lifeCycle,
     configuration,
+    location,
     session,
+    locationChangeObservable,
     () => ({
       context: {},
       user: {},
@@ -42,9 +47,12 @@ function startRum(
     configuration,
     location,
     domMutationObservable,
+    locationChangeObservable,
     foregroundContexts,
     noopRecorderApi
   )
+
+  startLongTaskCollection(lifeCycle)
   return {
     stop: () => {
       rumEventCollectionStop()
@@ -63,9 +71,25 @@ describe('rum session', () => {
     }
 
     setupBuilder = setup().beforeBuild(
-      ({ applicationId, location, lifeCycle, configuration, session, domMutationObservable }) => {
+      ({
+        applicationId,
+        location,
+        lifeCycle,
+        configuration,
+        session,
+        domMutationObservable,
+        locationChangeObservable,
+      }) => {
         serverRumEvents = collectServerEvents(lifeCycle)
-        return startRum(applicationId, lifeCycle, configuration, session, location, domMutationObservable)
+        return startRum(
+          applicationId,
+          lifeCycle,
+          configuration,
+          session,
+          location,
+          domMutationObservable,
+          locationChangeObservable
+        )
       }
     )
   })
@@ -107,10 +131,28 @@ describe('rum session keep alive', () => {
     setupBuilder = setup()
       .withFakeClock()
       .withSession(session)
-      .beforeBuild(({ applicationId, location, lifeCycle, configuration, session, domMutationObservable }) => {
-        serverRumEvents = collectServerEvents(lifeCycle)
-        return startRum(applicationId, lifeCycle, configuration, session, location, domMutationObservable)
-      })
+      .beforeBuild(
+        ({
+          applicationId,
+          location,
+          lifeCycle,
+          configuration,
+          session,
+          domMutationObservable,
+          locationChangeObservable,
+        }) => {
+          serverRumEvents = collectServerEvents(lifeCycle)
+          return startRum(
+            applicationId,
+            lifeCycle,
+            configuration,
+            session,
+            location,
+            domMutationObservable,
+            locationChangeObservable
+          )
+        }
+      )
   })
 
   afterEach(() => {
@@ -153,7 +195,7 @@ describe('rum session keep alive', () => {
   })
 })
 
-describe('rum view url', () => {
+describe('rum events url', () => {
   const FAKE_NAVIGATION_ENTRY: RumPerformanceNavigationTiming = {
     domComplete: 456 as RelativeTime,
     domContentLoadedEventEnd: 345 as RelativeTime,
@@ -168,9 +210,25 @@ describe('rum view url', () => {
 
   beforeEach(() => {
     setupBuilder = setup().beforeBuild(
-      ({ applicationId, location, lifeCycle, configuration, session, domMutationObservable }) => {
+      ({
+        applicationId,
+        location,
+        lifeCycle,
+        configuration,
+        session,
+        domMutationObservable,
+        locationChangeObservable,
+      }) => {
         serverRumEvents = collectServerEvents(lifeCycle)
-        return startRum(applicationId, lifeCycle, configuration, session, location, domMutationObservable)
+        return startRum(
+          applicationId,
+          lifeCycle,
+          configuration,
+          session,
+          location,
+          domMutationObservable,
+          locationChangeObservable
+        )
       }
     )
   })
@@ -179,12 +237,40 @@ describe('rum view url', () => {
     setupBuilder.cleanup()
   })
 
+  it('should attach the url corresponding to the start of the event', () => {
+    const { lifeCycle, clock, changeLocation } = setupBuilder
+      .withFakeClock()
+      .withFakeLocation('http://foo.com/')
+      .build()
+    clock.tick(10)
+    changeLocation('http://foo.com/?bar=bar')
+    clock.tick(10)
+    changeLocation('http://foo.com/?bar=qux')
+
+    lifeCycle.notify(LifeCycleEventType.PERFORMANCE_ENTRY_COLLECTED, {
+      entryType: 'longtask',
+      startTime: relativeNow() - 5,
+      toJSON: noop,
+      duration: 5,
+    } as RumPerformanceEntry)
+
+    clock.tick(THROTTLE_VIEW_UPDATE_PERIOD)
+
+    expect(serverRumEvents.length).toBe(3)
+    const [firstViewUpdate, longTaskEvent, lastViewUpdate] = serverRumEvents
+
+    expect(firstViewUpdate.view.url).toBe('http://foo.com/')
+    expect(lastViewUpdate.view.url).toBe('http://foo.com/')
+
+    expect(longTaskEvent.view.url).toBe('http://foo.com/?bar=bar')
+  })
+
   it('should keep the same URL when updating a view ended by a URL change', () => {
-    setupBuilder.withFakeLocation('http://foo.com/').build()
+    const { changeLocation } = setupBuilder.withFakeLocation('http://foo.com/').build()
 
     serverRumEvents.length = 0
 
-    history.pushState({}, '', '/bar')
+    changeLocation('/bar')
 
     expect(serverRumEvents.length).toEqual(2)
     expect(serverRumEvents[0].view.url).toEqual('http://foo.com/')
@@ -192,11 +278,14 @@ describe('rum view url', () => {
   })
 
   it('should keep the same URL when updating an ended view', () => {
-    const { lifeCycle, clock } = setupBuilder.withFakeClock().withFakeLocation('http://foo.com/').build()
+    const { lifeCycle, clock, changeLocation } = setupBuilder
+      .withFakeClock()
+      .withFakeLocation('http://foo.com/')
+      .build()
 
     clock.tick(VIEW_DURATION)
 
-    history.pushState({}, '', '/bar')
+    changeLocation('/bar')
 
     serverRumEvents.length = 0