+++ packages/rum-core/src/domain/rumEventsCollection/view/trackViews.spec.ts
@@ -45,91 +45,26 @@ describe('track views automatically', () => {
   })
 
   describe('location changes', () => {
-    it('should update view location on search change', () => {
-      setupBuilder.build()
-      const { getViewCreateCount, getViewCreate, getViewUpdate, getViewUpdateCount } = viewTest
-
-      history.pushState({}, '', '/foo?bar=qux')
-
-      expect(getViewCreateCount()).toBe(1)
-      expect(getViewCreate(0).location.href).toMatch(/\/foo$/)
-
-      const lastUpdate = getViewUpdate(getViewUpdateCount() - 1)
-      expect(lastUpdate.location.href).toMatch(/\/foo\?bar=qux$/)
-      expect(lastUpdate.id).toBe(getViewCreate(0).id)
-    })
-
     it('should create new view on path change', () => {
-      setupBuilder.build()
-      const { getViewCreateCount, getViewCreate } = viewTest
+      const { changeLocation } = setupBuilder.build()
+      const { getViewCreateCount } = viewTest
 
       expect(getViewCreateCount()).toBe(1)
-      expect(getViewCreate(0).location.href).toMatch(/\/foo$/)
 
-      history.pushState({}, '', '/bar')
+      changeLocation('/bar')
 
       expect(getViewCreateCount()).toBe(2)
-      expect(getViewCreate(1).location.href).toMatch(/\/bar$/)
     })
 
     it('should create new view on hash change from history', () => {
-      setupBuilder.build()
-      const { getViewCreateCount, getViewCreate } = viewTest
+      const { changeLocation } = setupBuilder.build()
+      const { getViewCreateCount } = viewTest
 
       expect(getViewCreateCount()).toBe(1)
-      expect(getViewCreate(0).location.href).toMatch(/\/foo$/)
 
-      history.pushState({}, '', '/foo#bar')
+      changeLocation('/foo#bar')
 
       expect(getViewCreateCount()).toBe(2)
-      expect(getViewCreate(1).location.href).toMatch(/\/foo#bar$/)
-    })
-
-    it('should not create a new view on hash change from history when the hash has kept the same value', () => {
-      history.pushState({}, '', '/foo#bar')
-
-      setupBuilder.build()
-      const { getViewCreateCount } = viewTest
-
-      expect(getViewCreateCount()).toBe(1)
-
-      history.pushState({}, '', '/foo#bar')
-
-      expect(getViewCreateCount()).toBe(1)
-    })
-
-    it('should create a new view on hash change', (done) => {
-      setupBuilder.build()
-      const { getViewCreateCount } = viewTest
-
-      function hashchangeCallBack() {
-        expect(getViewCreateCount()).toBe(2)
-        window.removeEventListener('hashchange', hashchangeCallBack)
-        done()
-      }
-
-      window.addEventListener('hashchange', hashchangeCallBack)
-
-      expect(getViewCreateCount()).toBe(1)
-      window.location.hash = '#bar'
-    })
-
-    it('should not create a new view when the hash has kept the same value', (done) => {
-      history.pushState({}, '', '/foo#bar')
-
-      setupBuilder.build()
-      const { getViewCreateCount } = viewTest
-
-      function hashchangeCallBack() {
-        expect(getViewCreateCount()).toBe(1)
-        window.removeEventListener('hashchange', hashchangeCallBack)
-        done()
-      }
-
-      window.addEventListener('hashchange', hashchangeCallBack)
-
-      expect(getViewCreateCount()).toBe(1)
-      window.location.hash = '#bar'
     })
 
     function mockGetElementById() {
@@ -137,154 +72,29 @@ describe('track views automatically', () => {
       return spyOn(document, 'getElementById').and.callFake(fakeGetElementById)
     }
 
-    it('should not create a new view when it is an Anchor navigation', (done) => {
-      setupBuilder.build()
+    it('should not create a new view when it is an Anchor navigation', () => {
+      const { changeLocation } = setupBuilder.build()
       const { getViewCreateCount } = viewTest
       mockGetElementById()
+      expect(getViewCreateCount()).toBe(1)
 
-      function hashchangeCallBack() {
-        expect(getViewCreateCount()).toBe(1)
-        window.removeEventListener('hashchange', hashchangeCallBack)
-        done()
-      }
-
-      window.addEventListener('hashchange', hashchangeCallBack)
+      changeLocation('/foo#testHashValue')
 
       expect(getViewCreateCount()).toBe(1)
-      window.location.hash = '#testHashValue'
     })
 
     it('should not create a new view when the search part of the hash changes', () => {
-      history.pushState({}, '', '/foo#bar')
-      setupBuilder.build()
+      const { changeLocation } = setupBuilder.build()
       const { getViewCreateCount } = viewTest
+      changeLocation('/foo#bar')
+      expect(getViewCreateCount()).toBe(2)
 
-      expect(getViewCreateCount()).toBe(1)
+      changeLocation('/foo#bar?search=1')
+      changeLocation('/foo#bar?search=2')
+      changeLocation('/foo#bar?')
+      changeLocation('/foo#bar')
 
-      history.pushState({}, '', '/foo#bar?search=1')
-      history.pushState({}, '', '/foo#bar?search=2')
-      history.pushState({}, '', '/foo#bar?')
-      history.pushState({}, '', '/foo#bar')
-
-      expect(getViewCreateCount()).toBe(1)
-    })
-  })
-
-  describe('view referrer', () => {
-    it('should set the document referrer as referrer for the initial view', () => {
-      setupBuilder.build()
-      const { getViewCreate } = viewTest
-
-      expect(getViewCreate(0).referrer).toEqual(document.referrer)
-    })
-
-    it('should set the previous view URL as referrer when a route change occurs', () => {
-      setupBuilder.build()
-      const { getViewCreate } = viewTest
-
-      history.pushState({}, '', '/bar')
-
-      expect(getViewCreate(1).referrer).toEqual(jasmine.stringMatching(/\/foo$/))
-    })
-
-    it('should set the previous view URL as referrer when a the session is renewed', () => {
-      const { lifeCycle } = setupBuilder.build()
-      const { getViewCreate } = viewTest
-
-      lifeCycle.notify(LifeCycleEventType.SESSION_RENEWED)
-
-      expect(getViewCreate(1).referrer).toEqual(jasmine.stringMatching(/\/foo$/))
-    })
-
-    it('should use the most up-to-date URL of the previous view as a referrer', () => {
-      setupBuilder.build()
-      const { getViewCreate } = viewTest
-
-      history.pushState({}, '', '/foo?a=b')
-      history.pushState({}, '', '/bar')
-
-      expect(getViewCreate(1).referrer).toEqual(jasmine.stringMatching(/\/foo\?a=b$/))
-    })
-  })
-})
-
-describe('track views manually', () => {
-  let setupBuilder: TestSetupBuilder
-  let viewTest: ViewTest
-
-  beforeEach(() => {
-    setupBuilder = setup()
-      .withFakeLocation('/foo')
-      .withConfiguration({ trackViewsManually: true })
-      .beforeBuild((buildContext) => {
-        viewTest = setupViewTest(buildContext)
-        return viewTest
-      })
-  })
-
-  afterEach(() => {
-    setupBuilder.cleanup()
-  })
-
-  describe('location changes', () => {
-    it('should update view location on search change', () => {
-      setupBuilder.build()
-      const { getViewCreateCount, getViewCreate, getViewUpdate, getViewUpdateCount } = viewTest
-
-      history.pushState({}, '', '/foo?bar=qux')
-
-      expect(getViewCreateCount()).toBe(1)
-      expect(getViewCreate(0).location.href).toMatch(/\/foo$/)
-
-      const lastUpdate = getViewUpdate(getViewUpdateCount() - 1)
-      expect(lastUpdate.location.href).toMatch(/\/foo\?bar=qux$/)
-      expect(lastUpdate.id).toBe(getViewCreate(0).id)
-    })
-
-    it('should update view location on path change', () => {
-      setupBuilder.build()
-      const { getViewCreateCount, getViewCreate, getViewUpdate, getViewUpdateCount } = viewTest
-
-      history.pushState({}, '', '/bar')
-
-      expect(getViewCreateCount()).toBe(1)
-      expect(getViewCreate(0).location.href).toMatch(/\/foo$/)
-
-      const lastUpdate = getViewUpdate(getViewUpdateCount() - 1)
-      expect(lastUpdate.location.href).toMatch(/\/bar$/)
-      expect(lastUpdate.id).toBe(getViewCreate(0).id)
-    })
-  })
-
-  describe('view referrer', () => {
-    it('should set the document referrer as referrer for the initial view', () => {
-      setupBuilder.build()
-      const { getViewCreate } = viewTest
-
-      expect(getViewCreate(0).referrer).toEqual(document.referrer)
-    })
-
-    it('should set the previous view URL as referrer when starting a new view', () => {
-      setupBuilder.build()
-      const { getViewUpdate, getViewUpdateCount, startView } = viewTest
-
-      startView()
-      history.pushState({}, '', '/bar')
-
-      const lastUpdate = getViewUpdate(getViewUpdateCount() - 1)
-      expect(lastUpdate.referrer).toEqual(jasmine.stringMatching(/\/foo$/))
-      expect(lastUpdate.location.href).toEqual(jasmine.stringMatching(/\/bar$/))
-    })
-
-    it('should use the most up-to-date URL of the previous view as a referrer', () => {
-      setupBuilder.build()
-      const { getViewCreate, startView } = viewTest
-
-      history.pushState({}, '', '/foo?a=b')
-      history.pushState({}, '', '/bar')
-      startView()
-
-      expect(getViewCreate(1).referrer).toEqual(jasmine.stringMatching(/\/bar$/))
+      expect(getViewCreateCount()).toBe(2)
     })
   })
 })
@@ -455,7 +265,7 @@ describe('renew session', () => {
   })
 
   it('should use the current view name for the new view', () => {
-    const { lifeCycle } = setupBuilder.build()
+    const { lifeCycle, changeLocation } = setupBuilder.build()
     const { getViewCreateCount, getViewCreate, startView } = viewTest
 
     lifeCycle.notify(LifeCycleEventType.SESSION_RENEWED)
@@ -465,7 +275,7 @@ describe('renew session', () => {
     lifeCycle.notify(LifeCycleEventType.SESSION_RENEWED)
 
     startView('qux')
-    history.pushState({}, '', '/bar')
+    changeLocation('/bar')
     lifeCycle.notify(LifeCycleEventType.SESSION_RENEWED)
 
     expect(getViewCreateCount()).toBe(8)