+++ test/e2e/lib/framework/serverApps/mock.ts
@@ -1,9 +1,10 @@
 import * as url from 'url'
 import cors from 'cors'
 import express from 'express'
-import { buildLogs, buildNpm, buildRum, buildRumSlim, Endpoints } from '../sdkBuilds'
+import { buildLogs, buildNpm, buildRum, buildRumSlim } from '../sdkBuilds'
+import { Servers } from '../httpServers'
 
-export function createMockServerApp(endpoints: Endpoints, setup: string) {
+export function createMockServerApp(servers: Servers, setup: string) {
   const app = express()
 
   app.use(cors())
@@ -47,24 +48,32 @@ export function createMockServerApp(endpoints: Endpoints, setup: string) {
   })
 
   app.get('/', (_req, res) => {
+    res.header(
+      'Content-Security-Policy',
+      [
+        `connect-src ${servers.intake.url} ${servers.base.url} ${servers.crossOrigin.url}`,
+        "script-src 'self' 'unsafe-inline'",
+        'worker-src blob:',
+      ].join(';')
+    )
     res.send(setup)
     res.end()
   })
 
   app.get('/datadog-logs.js', async (_req, res) => {
-    res.header('content-type', 'application/javascript').send(await buildLogs(endpoints))
+    res.header('content-type', 'application/javascript').send(await buildLogs(servers.intake.url))
   })
 
   app.get('/datadog-rum.js', async (_req, res) => {
-    res.header('content-type', 'application/javascript').send(await buildRum(endpoints))
+    res.header('content-type', 'application/javascript').send(await buildRum(servers.intake.url))
   })
 
   app.get('/datadog-rum-slim.js', async (_req, res) => {
-    res.header('content-type', 'application/javascript').send(await buildRumSlim(endpoints))
+    res.header('content-type', 'application/javascript').send(await buildRumSlim(servers.intake.url))
   })
 
   app.get('/app.js', async (_req, res) => {
-    res.header('content-type', 'application/javascript').send(await buildNpm(endpoints))
+    res.header('content-type', 'application/javascript').send(await buildNpm(servers.intake.url))
   })
 
   return app